add namespace rate limits
